Fix to check lowest level permission
When checking a permission, currently there is a bug where its parents permissions are prioritized over its own setting.

This fix checks the target permission first, and if none is set, checks the parents upward until one is found.

If no permissions anywhere are set, it uses the existing behavior of returning null.